re-structure quarantine screens - move to separate folder - split calendar view and add dates view - show add dates in a modal screen (stack nav) instead of toggle like view (since we didn;t have a way to go back)
