+++ go-corona-app/App.js
@@ -12,6 +12,7 @@ import { byPassLogin } from './constants/DevSettings';
 import RootView from './navigation/BottomTabNavigator';
 import useLinking from './navigation/useLinking';
 import LoginScreen from './screens/LoginScreen';
+import QuarantineDates from './screens/quarantine/QuarantineDates';
 import TravelQuestionnaire from "./screens/crosscheck/travel/TravelScreen";
 import IntersectionCalculator from './screens/crosscheck/intersection/index';
 import { Provider as PaperProvider } from 'react-native-paper';
@@ -63,6 +64,7 @@ export default function App(props) {
           <AppStack.Navigator initialRouteName={byPassLogin ? "Root" : "Login"}>
             <AppStack.Screen name="Login" component={LoginScreen} />
             <AppStack.Screen name="Root" component={RootView} />
+            <AppStack.Screen name="QuarantineDates" options={{ title: "Edit"}} component={QuarantineDates} />
             <AppStack.Screen name="TravelQuestionnaire" options={{ title: "Travel Questionnaire"}} component={TravelQuestionnaire} />
             <AppStack.Screen name="IntersectionCalculator" options={{ title: "Intersection Calculator"}} component={IntersectionCalculator} />
           </AppStack.Navigator>
