+++ go-corona-app/navigation/BottomTabNavigator.js
@@ -4,7 +4,7 @@ import { View, Text } from 'react-native';
 
 import TabBarIcon, {MaterialTabBarIcon} from '../components/TabBarIcon';
 import HomeScreen from '../screens/HomeScreen';
-import QuarantineScreen from '../screens/QuarantineScreen';
+import QuarantineScreen from '../screens/quarantine/QuarantineScreen';
 import CheckupScreen from '../screens/checkup/CheckupScreen';
 import CrosscheckScreen from '../screens/crosscheck/CrosscheckScreen';
 import InfoScreen from '../screens/InfoScreen';
